Possible warning fixes?

+++ src/sfizz/Synth.cpp
@@ -426,7 +426,7 @@ void Synth::Impl::handleControlOpcodes(const std::vector<Opcode>& members)
 	    {
 		if (int(stoi(member.value) / sampleRate_) > 1)
 			overSampled = true;
-		resources_.synthConfig.OSFactor = std::min(128, std::max(1, int(stoi(member.value) / sampleRate_)));
+		resources_.synthConfig.OSFactor = int(std::min(128, std::max(1, int(stoi(member.value) / sampleRate_))));
     		for (auto& voice : voiceManager_) {
         		voice.setSampleRate(sampleRate_);
         		voice.setSamplesPerBlock(samplesPerBlock_);
